FINAL_resources/app_demo/field/requirements.txt
field==0.2.0
matplotlib==3.10.8
numpy==2.3.5
opencv_python==4.11.0.86
pandas==2.3.3
Pillow==12.0.0
scikit_learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
timm==1.0.22
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
